refactor slug-related stuff out of service.rb and into a concern
